discord_finance_bot/requirements.txt
discord.py
apscheduler
yfinance
requests>=2.32.0
beautifulsoup4>=4.12.3
python-dotenv>=1.0.1
playwright>=1.47.0